cleaned up docstrings
