--- .gitignore
@@ -1,15 +1,0 @@
-# Compiled Bytes
-__pycache__/
-*.py[cod]
-
-# VIM Temporary Files
-.*.swp
-
-# Jupyter Notebook Checkpoints
-.ipynb_checkpoints
-
-# Virtualenv
-env/
-
-# Bazel Folders and Symlinks
-bazel-*
