+++ examples/mnist_v2/BUILD
