+++ examples/mnist_v2/README.md
