+++ examples/mnist_v2/export_test.py
@@ -2,15 +2,12 @@ import tensorflow as tf
 import tensorflow_hub as hub
 import unittest
 import tempfile
-import shutil
 import os
 from absl.testing import absltest
 import export
-from export import train_and_export
 
-TMPDIR = tempfile.mkdtemp()
 model = None
-TEST_DATA_DIR = "examples/mnist/test_data"
+TEST_DATA_DIR = "examples/mnist_v2/test_data"
 
 def load_model(path):
   global model
@@ -35,30 +32,26 @@ class TFHubMNISTTest(tf.test.TestCase):
           export.sys.__dict__,
           mocked_dict,
           clear=True).start()
-      train_and_export(epoch=1, export_path="%s/model/1" % TMPDIR)
-      self.assertTrue(os.listdir(TMPDIR))
+      export.train_and_export(epoch=1, export_path="%s/model/1" % self.get_temp_dir())
+      self.assertTrue(os.listdir(self.get_temp_dir()))
 
   def test_empty_input(self):
-    if not os.path.exists("%s/model/1" % TMPDIR):
+    if not os.path.exists("%s/model/1" % self.get_temp_dir()):
       self.test_model_exporting()
-    model = load_model("%s/model/1" % TMPDIR)
+    model = load_model("%s/model/1" % self.get_temp_dir())
     output_ = model.call(
         tf.zeros([1, 28, 28, 1], dtype=tf.uint8).numpy())
     self.assertEqual(output_.shape, [1, 10])
 
   def test_sample_input(self):
-    if not os.path.exists(os.path.join(TMPDIR, "model/1")):
+    if not os.path.exists(os.path.join(self.get_temp_dir(), "model/1")):
       self.test_model_exporting()
-    model = load_model("%s/model/1" % TMPDIR)
+    model = load_model("%s/model/1" % self.get_temp_dir())
     prediction = model.call(self.test_image)
     self.assertEqual(
         tf.argmax(
             tf.squeeze(prediction)).numpy(),
         self.test_label)
-
-  @classmethod
-  def tearDownClass(cls):
-    shutil.rmtree(TMPDIR)
 
 if __name__ == '__main__':
   tf.test.main()
