+++ examples/mnist_v2/test_data/8.jpg
